# src/main/resources/swagger/provider-bs-service.yml
---
swagger: "2.0"
info:
  description: "RESTful API designed as the Provider Registry of the Federated ebox\
    \ for citizens."
  version: "0.5"
  title: "Ebox Citizen provider registry REST service"
  contact:
    name: "Ebox Development Team"
    url: "http://www.smals.be"
    email: "citizenfedbox@smals.be"
  license:
    name: "digitalbelgium"
    url: "http://digitalbelgium.be/"
basePath: "/"
tags:
  - name: "Documentation"
  - name: "MessageProviders"
  - name: "Monitoring"
schemes:
  - "https"
paths:
  /providerregistry:
    get:
      tags:
        - "MessageProviders"
      summary: "List all message providers. Message providers are services that implement\
        \ the REST API to allow user interfaces to request messages for citizens."
      description: ""
      operationId: "findAllMessageProviders"
      produces:
        - "application/json"
      parameters:
        - name: "page"
          in: "query"
          description: "The index of the current page of items, default 1"
          required: false
          type: "integer"
          default: 1
          minimum: 1.0
          format: "int32"
        - name: "pageSize"
          in: "query"
          description: "The maximal number of results per page, default 25"
          required: false
          type: "integer"
          default: 25
          minimum: 1.0
          format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/MessageProviders"
        404:
          description: "MessageProviders not found"
          schema:
            $ref: "#/definitions/StatusMessage"
  /providerregistry/provider-api:
    get:
      tags:
        - "Documentation"
      summary: "List links"
      description: ""
      operationId: "get"
      produces:
        - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/HalResource"
          headers:
            Link:
              type: "string"
              description: "RFC 5988 Link Header"
    options:
      tags:
        - "Documentation"
      summary: "List links"
      description: ""
      operationId: "options"
      produces:
        - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/HalResource"
          headers:
            Link:
              type: "string"
              description: "RFC 5988 Link Header"
  /providerregistry/provider-api/doc:
    get:
      tags:
        - "Documentation"
      summary: "Returns the API documentation in HTML format"
      description: ""
      operationId: "getApiDoc"
      produces:
        - "text/html"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
  /providerregistry/provider-api/doc/jsonSchemas/{jsonSchema}:
    get:
      tags:
        - "Documentation"
      summary: "Returns the JSON Schemas"
      description: ""
      operationId: "getJsonSchema"
      produces:
        - "application/json"
      parameters:
        - name: "jsonSchema"
          in: "path"
          description: "JSON Schema name"
          required: true
          type: "string"
          pattern: ".+"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
  /providerregistry/provider-api/doc/swagger.json:
    get:
      tags:
        - "Documentation"
      summary: "Returns the API documentation in JSON Swagger format"
      description: ""
      operationId: "getSwagger"
      produces:
        - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
  /providerregistry/provider-api/healthCheck:
    get:
      tags:
        - "Monitoring"
      summary: "Check health of this REST API"
      description: "This resource is only available to supervision users"
      operationId: "healthCheck"
      produces:
        - "application/json"
        - "text/xml"
      parameters:
        - name: "type"
          in: "query"
          description: "Defines the which sanity checks must be ran"
          required: false
          type: "string"
          default: "default"
          enum:
            - "ping"
            - "default"
            - "deep"
      responses:
        default:
          description: "successful operation"
  /providerregistry/{messageProviderId}:
    get:
      tags:
        - "MessageProviders"
      summary: "Get a messageProvider by its messageProviderId"
      description: ""
      operationId: "getMessageProviderById"
      produces:
        - "application/json"
      parameters:
        - name: "messageProviderId"
          in: "path"
          description: "The messageProviderId of the provider"
          required: true
          type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/MessageProvider"
        400:
          description: "Invalid messageProviderId supplied"
          schema:
            $ref: "#/definitions/StatusMessage"
        404:
          description: "MessageProviderId not found"
          schema:
            $ref: "#/definitions/StatusMessage"
  /providerregistry/{messageProviderId}/images/{imageId}/{formatId}/content:
    get:
      tags:
        - "MessageProviders"
      summary: "Get an image by its imageId"
      description: ""
      operationId: "getImageById"
      produces:
        - "application/octet-stream"
      parameters:
        - name: "imageId"
          in: "path"
          description: "The identifier of the image"
          required: true
          type: "string"
        - name: "messageProviderId"
          in: "path"
          description: "The identifier of the institution"
          required: true
          type: "string"
        - name: "formatId"
          in: "path"
          description: "The size of the image"
          required: true
          type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
            format: "binary"
        400:
          description: "Invalid imageId supplied"
          schema:
            $ref: "#/definitions/StatusMessage"
        404:
          description: "Image not found"
          schema:
            $ref: "#/definitions/StatusMessage"
definitions:
  Detail:
    type: "object"
    properties:
      kind:
        type: "object"
      message:
        type: "object"
      ref:
        type: "object"
      value:
        type: "object"
  HalLink:
    type: "object"
    required:
      - "href"
    properties:
      href:
        type: "string"
        format: "uri"
        description: "Its value is either a URI [RFC3986] or a URI Template [RFC6570]."
      hreflang:
        type: "string"
        description: "When present, is a hint in RFC5646 format indicating what the\
          \ language of the result of dereferencing the link should be.  Note that\
          \ this is only a hint; for example, it does not override the Content-Language\
          \ header of a HTTP response obtained by actually following the link."
        pattern: "^([a-zA-Z]{2,3}(-[a-zA-Z]{3}(-[a-zA-Z]{3}){0,2})?(-[a-zA-Z]{4})?(-([a-zA-Z]{2}|[0-9]{3}))?(-([a-zA-Z0-9]{5,8}|[0-9][a-zA-Z0-9]{3}))*([0-9A-WY-Za-wy-z](-[a-zA-Z0-9]{2,8}){1,})*(x-[a-zA-Z0-9]{2,8})?)|(x-[a-zA-Z0-9]{2,8})|(en-GB-oed)|(i-ami)|(i-bnn)|(i-default)|(i-enochian)|(i-hak)|(i-klingon)|(i-lux)|(i-mingo)|(i-navajo)|(i-pwn)|(i-tao)|(i-tay)|(i-tsu)|(sgn-BE-FR)|(sgn-BE-NL)|(sgn-CH-DE)|(art-lojban)|(cel-gaulish)|(no-bok)|(no-nyn)|(zh-guoyu)|(zh-hakka)|(zh-min)|(zh-min-nan)|(zh-xiang)$"
      templated:
        type: "boolean"
        description: "Its value is boolean and SHOULD be true when the Link Object's\
          \ href property is a URI Template."
        default: false
      type:
        type: "string"
        description: "Its value is a string used as a hint to indicate the media type\
          \ expected when dereferencing the target resource."
      deprecation:
        type: "boolean"
        description: "Its presence indicates that the link is to be deprecated (i.e.\
          \ removed) at a future date. Its value is a URL that SHOULD provide further\
          \ information about the deprecation."
        default: false
      name:
        type: "string"
        description: "Its value MAY be used as a secondary key for selecting Link\
          \ Objects which share the same relation type."
      title:
        type: "string"
        description: "Its value is a string and is intended for labelling the link\
          \ with a human-readable identifier (as defined by [RFC5988])."
  HalLinks:
    type: "object"
    properties:
      self:
        $ref: "#/definitions/HalLink"
    description: "Object of links with the rels as the keys"
  HalResource:
    type: "object"
    properties:
      _links:
        description: "Object of links with the rels as the keys"
        $ref: "#/definitions/HalLinks"
    description: "A resource modeled after the HAL specification"
  Image:
    type: "object"
    required:
      - "contentHref"
      - "format"
      - "size"
    properties:
      imageId:
        type: "string"
        description: "Identifier of the image"
      format:
        type: "string"
        description: "Size of the image defined in one of the available sizes: small\
          \ (less then 100px), medium (between 100px and 500px) and large (more than\
          \ 500px)"
        enum:
          - "small"
          - "medium"
          - "large"
      size:
        type: "number"
        format: "double"
        example: 1800.0
        description: "Size of the image, expressed in kB"
      mediaType:
        type: "string"
        description: "Format of the image following RFC2046"
        enum:
          - "image/jpg"
          - "image/jpeg"
          - "image/png"
          - "image/gif"
          - "image/bmp"
          - "image/svg+xml"
      language:
        type: "string"
        description: "Language of the image"
      contentHref:
        type: "string"
        format: "uri"
      _links:
        description: "Object of links with the rels as the keys"
        $ref: "#/definitions/HalLinks"
    description: "An image gives the metadata and the base64 content of the image"
  MessageProvider:
    type: "object"
    required:
      - "messageProviderActivationDate"
      - "messageProviderDescription"
      - "messageProviderId"
      - "messageProviderLogo"
      - "messageProviderName"
      - "messageProviderPath"
      - "messageProviderPlannedMaintenance"
      - "messageProviderTechnicalContact"
      - "messageProviderUrl"
    properties:
      messageProviderId:
        type: "string"
        example: "eBoxCitizen"
        description: "Identifier of the messageProvider"
        minLength: 1
        maxLength: 100
      messageProviderName:
        description: "Common type defined to provide a string value for each supported\
          \ language among nl, fr, de, en"
        $ref: "#/definitions/TranslatedString"
      messageProviderDescription:
        description: "Common type defined to provide a string value for each supported\
          \ language among nl, fr, de, en"
        $ref: "#/definitions/TranslatedString"
      messageProviderLogo:
        type: "array"
        description: "Logo of the provider"
        items:
          $ref: "#/definitions/Image"
      messageProviderUrl:
        type: "string"
        example: "http://ebox-citizen-rest.int.paas.socialsecurity.be"
        description: "The base URL to the messageProvider implementation"
      messageProviderPath:
        type: "string"
        example: "/REST/ebox/citizen/v2/"
        description: "The path on which the messageProvider is available on the URL"
      messageProviderActivationDate:
        type: "string"
        format: "date"
        example: "2017-12-01"
        description: "Activation date of the service. Services should be ignored when\
          \ they are not yet activated."
      messageProviderDeactivationDate:
        type: "string"
        format: "date"
        example: "2018-01-01"
        description: "Deactivation date of the service. Services should be ignored\
          \ when they are deactivated."
      messageProviderTechnicalContact:
        type: "string"
        example: "eboxintegration@smals.be"
        description: "Contact information for technical questions concerning the implementation\
          \ of the REST service by the provider"
      messageProviderPlannedMaintenance:
        type: "array"
        description: "List of dateIntervals on which the REST service of the provider\
          \ will be unavailable due to maintenance"
        items:
          $ref: "#/definitions/MessageProviderPlannedMaintenance"
      _links:
        description: "Object of links with the rels as the keys"
        $ref: "#/definitions/HalLinks"
    description: "Schema describing the messageProvider resource that returns the\
      \ necessary information concerning a messageResource"
  MessageProviderPlannedMaintenance:
    type: "object"
    required:
      - "endDate"
      - "startDate"
    properties:
      startDate:
        type: "string"
        format: "date-time"
        example: "2017-09-09T08:00:00Z"
        description: "Date and time from which the service will no longer be available\
          \ because of maintenance."
      endDate:
        type: "string"
        format: "date-time"
        example: "2017-09-09T20:00:00Z"
        description: "Date and time from which the service will be available again."
    description: "A date interval defined by a start date-time and an end date-time"
  MessageProviders:
    type: "object"
    properties:
      items:
        type: "array"
        items:
          $ref: "#/definitions/MessageProvider"
      totalItems:
        type: "integer"
        format: "int32"
        description: "The number of messageProviders in the array"
    description: "List of messageProviders"
  StatusMessage:
    type: "object"
    required:
      - "code"
      - "id"
      - "message"
    properties:
      id:
        type: "string"
        format: "uuid"
      code:
        type: "string"
      message:
        type: "string"
      contact:
        type: "string"
      environment:
        type: "string"
        enum:
          - "development"
          - "test"
          - "integration"
          - "acceptation"
          - "simulation"
          - "production"
      stackTrace:
        type: "array"
        items:
          type: "array"
          items:
            type: "string"
      details:
        type: "array"
        items:
          $ref: "#/definitions/Detail"
    description: "A representation of a generic error message"
  TranslatedString:
    type: "object"
    properties:
      nl:
        type: "string"
      fr:
        type: "string"
      de:
        type: "string"
      en:
        type: "string"
    description: "Common type defined to provide a string value for each supported\
      \ language among nl, fr, de, en"